refactor: remove multibase check
https://github.com/ipfs/is-ipfs/pull/20#discussion_r204333276
