+++ package.json
@@ -31,7 +31,6 @@
   "dependencies": {
     "bs58": "4.0.1",
     "cids": "0.5.3",
-    "multibase": "0.4.0",
     "multihashes": "0.4.13"
   },
   "devDependencies": {
